fix(domain): prizeMoney와 totalPrize를 Long으로 변경하여 오버플로우 방지
- Rank enum, LottoResult 클래스에서 prizeMoney, totalPrize 타입을 Int에서 Long으로 변경

+++ src/main/kotlin/lotto/domain/LottoResult.kt
@@ -3,7 +3,7 @@ package lotto.domain
 class LottoResult(private val ranks: Map<Rank, Int>) {
     fun getRanks(): Map<Rank, Int> = ranks
 
-    fun calculateTotalPrize(): Int{
+    fun calculateTotalPrize(): Long{
         return ranks.entries.sumOf { it.key.prizeMoney * it.value }
     }
